Remove unnecessary nesting of helper classes
No need to nest these in `*Helper` modules, and not doing so makes them
easier to find.
