+++ app/views/partials/_date_time_select.html.erb
@@ -1,5 +1,5 @@
 
-<% helper = DateTimeSelectHelper::DateTimeSelect.new(model, datetime_field_name) %>
+<% helper = DateTimeSelect.new(model, datetime_field_name) %>
 
 <%#
   This label is not very semantic since it just refers to the day field (as
